Modified Xamarin
Made small adjustments, addes ViewCartPage
