+++ CSE/Xamarin/Xamarin/GraphPage.xaml
@@ -1,18 +1,9 @@
 ﻿<?xml version="1.0" encoding="utf-8" ?>
 <ContentPage xmlns="http://xamarin.com/schemas/2014/forms"
              xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
-               xmlns:oxy="clr-namespace:OxyPlot.Xamarin.Forms;assembly=OxyPlot.Xamarin.Forms"
-             x:Class="Xamarin.GraphPage"
-             Padding="20">
-    <ContentPage.ToolbarItems>
-        <ToolbarItem x:Name="mainToolbar" Text="Main" Clicked="mainToolbar_Clicked"/>
-    </ContentPage.ToolbarItems>
+              xmlns:oxy="clr-namespace:OxyPlot.Xamarin.Forms;assembly=OxyPlot.Xamarin.Forms"
+             x:Class="Xamarin.GraphPage">
     <ContentPage.Content>
-        <StackLayout>
-            <Picker x:Name="picker" Title="Select a store" />
-            <ListView x:Name="list"/>
-            <Button x:Name="button" Text="View graph" Clicked="button_Clicked"/>
-            <oxy:PlotView Model="{Binding  model}" VerticalOptions="Center" HorizontalOptions="Center" />
-        </StackLayout>
+        <oxy:PlotView Model="{Binding  model}" VerticalOptions="Center" HorizontalOptions="Center" />
     </ContentPage.Content>
 </ContentPage>
