+++ CSE/Xamarin/Xamarin/GraphPage.xaml.cs
@@ -27,69 +27,81 @@ namespace Xamarin
             Norfa = 4
         }
 
-        const string path = "http://192.168.0.106:5000/api/"; //use your IP - command, ipconfig
-        string item;
-        string storeName;
+        string path = "http://192.168.0.106:5000/api/"; //use your IP - command, ipconfig
+        string item = "Pienas";
+        Store storeName = Store.IKI;
         LineSeries line;
-        List<String> productList;
         Dictionary<DateTime, decimal> listForDays;
         List<Dictionary<DateTime, decimal>> listOfLists = new List<Dictionary<DateTime, decimal>>();
 
-        public GraphPage()
+        public GraphPage(string it, string store)
         {
             InitializeComponent();
-            button.IsVisible = false;
-            button.IsEnabled = false;
-            List<Store> stores = Enum.GetValues(typeof(Store)).Cast<Store>().ToList();
-            foreach(var store in stores)
-            {
-                picker.Items.Add(store.ToString());
-            }
-
-            picker.SelectedIndexChanged += async (sender, args) =>
-            {
-                storeName = picker.SelectedItem.ToString();
-                //storeName = (Store)Enum.Parse(typeof(Store), selected);
-                await GetProductsData();
-                list.ItemsSource = productList;
-                list.ItemSelected += async (sender2, args2) =>
-                {
-                    item = list.SelectedItem.ToString();
-                    //picker.IsVisible = false;
-                    //list.IsVisible = false;
-                    button.IsVisible = true;
-                    button.IsEnabled = true;
-                };
-            };
+            item = it;
+            storeName = (Store)Enum.Parse(typeof(Store), store);
+            AddLine();
         }
-        
-        public async Task GetProductsData()
+
+        public async Task GetAPIData()
         {
             try
             {
-                string path1 = path + $"graph/{storeName.ToString()}";
-                
-                using (var client = new RestClient(new Uri(path1)))
+                path += "graph/?item=" + item + "&storeName=" + ((int)storeName).ToString();
+                var pathh = path;
+                using (var client = new RestClient(new Uri(path)))
                 {
                     var request = new RestRequest(Method.GET);
-                    var result = await client.Execute<IEnumerable<String>>(request);
-                    productList = result.Data.ToList();
+                    var result = await client.Execute<List<Dictionary<DateTime, decimal>>>(request);
+                    listOfLists = result.Data;
                 }
             }
             catch (Exception e)
             {
-                Debug.WriteLine("Error  == " + e.Message);
+                Debug.WriteLine("Error  == " + e.InnerException.Message);
             }
         }
 
-        async void mainToolbar_Clicked(object sender, EventArgs e)
+        public async Task GetList()
         {
-            await Navigation.PushAsync(new MainPage());
+            await GetAPIData();
+            listForDays = listOfLists[0];
+            var ordered = from pair in listForDays
+                          orderby pair.Key ascending
+                          select pair;
+            listForDays = ordered.ToDictionary(pair => pair.Key, pair => pair.Value);
         }
 
-        async void button_Clicked(object sender, EventArgs e)
+        public async void AddLine()
         {
-            await Navigation.PushAsync(new StatisticsPage(item, storeName));
+            await GetList();
+            var model = new PlotModel { Title = $"Product price change" };
+            Content = new PlotView
+            {
+                Model = model,
+                VerticalOptions = LayoutOptions.CenterAndExpand,
+                HorizontalOptions = LayoutOptions.CenterAndExpand,
+            };
+            line = new LineSeries();
+
+            DateTimeAxis xAxis = new DateTimeAxis
+            {
+                Position = AxisPosition.Bottom,
+                StringFormat = "yyyy/MM/dd",
+
+                Title = "Date",
+                MinorIntervalType = DateTimeIntervalType.Days,
+                IntervalType = DateTimeIntervalType.Days,
+                MajorGridlineStyle = LineStyle.Solid,
+                MinorGridlineStyle = LineStyle.None,
+            };
+            model.Axes.Add(xAxis);
+            model.Axes.Add(new LinearAxis { Title = "Price" });
+            foreach (var days in listForDays)
+            {
+                line.Points.Add(new DataPoint(DateTimeAxis.ToDouble(days.Key), Convert.ToDouble(days.Value)));
+            }
+            model.Series.Add(line);
         }
+        
     }
 }
