+++ CSE/Xamarin/Xamarin/StatisticsPage.xaml.cs
@@ -27,81 +27,65 @@ namespace Xamarin
             Norfa = 4
         }
 
-        string path = "http://192.168.0.106:5000/api/"; //use your IP - command, ipconfig
-        string item = "Pienas";
-        Store storeName = Store.IKI;
-        LineSeries line;
-        Dictionary<DateTime, decimal> listForDays;
+        const string path = "http://192.168.0.106:5000/api/"; //use your IP - command, ipconfig
+        string item;
+        string storeName;
+        List<String> productList;
         List<Dictionary<DateTime, decimal>> listOfLists = new List<Dictionary<DateTime, decimal>>();
 
-        public StatisticsPage(string it, string store)
+        public StatisticsPage()
         {
             InitializeComponent();
-            item = it;
-            storeName = (Store)Enum.Parse(typeof(Store), store);
-            AddLine();
-        }
+            Button.IsVisible = false;
+            Button.IsEnabled = false;
+            List<Store> stores = Enum.GetValues(typeof(Store)).Cast<Store>().ToList();
+            foreach(var store in stores)
+            {
+                picker.Items.Add(store.ToString());
+            }
 
-        public async Task GetAPIData()
+            picker.SelectedIndexChanged += async (sender, args) =>
+            {
+                storeName = picker.SelectedItem.ToString();
+                
+                await GetProductsData();
+                list.ItemsSource = productList;
+                list.ItemSelected += (sender2, args2) =>
+                {
+                    item = list.SelectedItem.ToString();
+                    Button.IsVisible = true;
+                    Button.IsEnabled = true;
+                };
+            };
+        }
+        
+        public async Task GetProductsData()
         {
             try
             {
-                path += "graph/?item=" + item + "&storeName=" + ((int)storeName).ToString();
-                var pathh = path;
-                using (var client = new RestClient(new Uri(path)))
+                string path1 = path + $"graph/{storeName.ToString()}";
+                
+                using (var client = new RestClient(new Uri(path1)))
                 {
                     var request = new RestRequest(Method.GET);
-                    var result = await client.Execute<List<Dictionary<DateTime, decimal>>>(request);
-                    listOfLists = result.Data;
+                    var result = await client.Execute<IEnumerable<String>>(request);
+                    productList = result.Data.ToList();
                 }
             }
             catch (Exception e)
             {
-                Debug.WriteLine("Error  == " + e.InnerException.Message);
+                Debug.WriteLine("Error  == " + e.Message);
             }
         }
 
-        public async Task GetList()
+        async void MainToolbar_Clicked(object sender, EventArgs e)
         {
-            await GetAPIData();
-            listForDays = listOfLists[0];
-            var ordered = from pair in listForDays
-                          orderby pair.Key ascending
-                          select pair;
-            listForDays = ordered.ToDictionary(pair => pair.Key, pair => pair.Value);
+            await Navigation.PushAsync(new MainPage());
         }
 
-        public async void AddLine()
+        async void Button_Clicked(object sender, EventArgs e)
         {
-            await GetList();
-            var model = new PlotModel { Title = $"Product price change" };
-            Content = new PlotView
-            {
-                Model = model,
-                VerticalOptions = LayoutOptions.CenterAndExpand,
-                HorizontalOptions = LayoutOptions.CenterAndExpand,
-            };
-            line = new LineSeries();
-
-            DateTimeAxis xAxis = new DateTimeAxis
-            {
-                Position = AxisPosition.Bottom,
-                StringFormat = "yyyy/MM/dd",
-
-                Title = "Date",
-                MinorIntervalType = DateTimeIntervalType.Days,
-                IntervalType = DateTimeIntervalType.Days,
-                MajorGridlineStyle = LineStyle.Solid,
-                MinorGridlineStyle = LineStyle.None,
-            };
-            model.Axes.Add(xAxis);
-            model.Axes.Add(new LinearAxis { Title = "Price" });
-            foreach (var days in listForDays)
-            {
-                line.Points.Add(new DataPoint(DateTimeAxis.ToDouble(days.Key), Convert.ToDouble(days.Value)));
-            }
-            model.Series.Add(line);
+            await Navigation.PushAsync(new GraphPage(item, storeName));
         }
-        
     }
 }
