+++ CSE/Xamarin/Xamarin/ViewCartPage.xaml
@@ -1,7 +1,7 @@
 ﻿<?xml version="1.0" encoding="utf-8" ?>
 <ContentPage xmlns="http://xamarin.com/schemas/2014/forms"
              xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
-             x:Class="Xamarin.CartPage"
+             x:Class="Xamarin.ViewCartPage"
              Padding="20">
     <ContentPage.ToolbarItems>
         <ToolbarItem x:Name="MainToolbar" Text="Main" Clicked="MainToolbar_Clicked"/>
@@ -9,11 +9,10 @@
     <ContentPage.Content>
         <StackLayout VerticalOptions="CenterAndExpand">
             <ScrollView>
-                <ListView x:Name="ListView"  ItemSelected="ListViewSelected" />
+                <ListView x:Name="ListView"/>
             </ScrollView>
-            <Button x:Name="AddItemButton" Text="Add Item" Clicked="AddItemButton_Clicked"/>
-            <Button x:Name="RemoveItemButton" Text="Remove Item" Clicked="RemoveItemButton_Clicked"/>
-            <Button x:Name="ViewCartButton" Text="View cart" Clicked="ViewCartButton_Clicked"/>
+            <Button x:Name="ResumeButton" Text="Resume choosing products" Clicked="ResumeButton_Clicked"/>
+            <Button x:Name="CheapestStoreButton" Text ="Show the cheapest store" Clicked="CheapestStoreButton_Clicked"/>
         </StackLayout>
     </ContentPage.Content>
 </ContentPage>
